requirements.txt
numpy==1.24.3
opencv_python==4.10.0.84
pandas==2.0.3
Pillow==10.4.0
scikit_learn==1.3.0
streamlit==1.38.0
tensorflow==2.10.0
